refactor: uniformizar ids relacionados do colaborador

+++ src/main/java/br/com/nord_tool_backend/domain/Colaborador.java
@@ -15,10 +15,10 @@ public class Colaborador extends GlobalDomain {
     private Long id;
     private String nome;
     private String celular;
-    private Integer idEmpresa;
+    private Long idEmpresa;
     private String nomeEmpresa;
-    private Integer idCargo;
+    private Long idCargo;
     private String nomeCargo;
-    private Integer idPermissao;
+    private Long idPermissao;
     private String nomePermissao;
 }
